converting to laz support

+++ little_awk/bedforms.py
@@ -11,15 +11,19 @@ S. Filhol, September 2022
 import matplotlib.pyplot as plt
 from matplotlib.colors import LightSource
 import pandas as pd
+import os
 
+def plot_meteorology(ts):
+    print('to be implemented')
+    return
 
-def plot_surface_change(ds, start, end, path, fname='test_*.png', figsize=(5,10), hillshade=True, dx=0.1, dy=0.1, **kwargs):
+def plot_surface_change(ds, start, end, path, fname='test_*.png', figsize=(10,5), hillshade=True, dx=0.1, dy=0.1, **kwargs):
     '''
     Function to plot frames for animation
     In Progress ...
 
     Args:
-        ds: 
+        ds:
         start:
         end:
         path:
@@ -33,30 +37,33 @@ def plot_surface_change(ds, start, end, path, fname='test_*.png', figsize=(5,10)
     Returns:
 
     '''
-    n_digits = str(ds.time.sel(time=slice(start, end)).shape).__len__() + 1
-    for i in ds.time.sel(time=slice(start, end)):
-        surf = ds.snow_surface.sel(time=i) - ds.snow_surface.sel(time=start)
+    if os.path.exists(path):
+        n_digits = str(ds.time.sel(time=slice(start, end)).shape).__len__() + 1
+        for i, date in enumerate(ds.time.sel(time=slice(start, end))):
+            surf = ds.snow_surface.sel(time=date) - ds.snow_surface.sel(time=start)
 
-        fig = plt.figure(figsize=figsize)
-        if hillshade:
-            ls = LightSource(azdeg=315, altdeg=45)
-            shade = ls.hillshade(surf.values, vert_exag=0.5,
-                                 dx=dx,
-                                 dy=dy,
-                                 fraction=1.0)
+            fig = plt.figure(figsize=figsize)
+            if hillshade:
+                ls = LightSource(azdeg=315, altdeg=45)
+                shade = ls.hillshade(surf.values, vert_exag=0.5,
+                                     dx=dx,
+                                     dy=dy,
+                                     fraction=1.0)
 
 
-            plt.imshow(shade,
-                       extent=[ds.x.min(), ds.x.max(), ds.y.min(), ds.y.max()],
-                       cmap=plt.cm.gray)
-        alpha=0.5
-        plt.imshow(surf.values,
-                       extent=[ds.x.min(), ds.x.max(), ds.y.min(), ds.y.max()],
-                       cmap=plt.cm.gray, **kwargs)
-        plt.colorbar()
-        plt.title(pd.to_datetime(ds.time.sel(time=i).values).strftime('%Y-%m-%d %H:%M'))
-        plt.savefig(path + fname.split('*')[0] + str(i).zfill(n_digits) + fname.split('*')[1])
-        plt.close()
+                plt.imshow(shade,
+                           extent=[ds.x.min(), ds.x.max(), ds.y.min(), ds.y.max()],
+                           cmap=plt.cm.gray)
+            alpha=0.5
+            plt.imshow(surf.values,
+                           extent=[ds.x.min(), ds.x.max(), ds.y.min(), ds.y.max()],
+                           cmap=plt.cm.viridis, alpha=.5, vmin=-0.1, vmax=0.5, **kwargs)
+            plt.colorbar()
+            plt.title(pd.to_datetime(ds.time.sel(time=date).values).strftime('%Y-%m-%d %H:%M'))
+            plt.savefig(path + fname.split('*')[0] + str(i).zfill(n_digits) + fname.split('*')[1])
+            plt.close()
 
-    cmd = 'ffmpeg to be coded'
-    print(f'run ffmpeg command: \n {cmd}\n')
+        cmd = 'ffmpeg to be coded'
+        print(f'run ffmpeg command: \n {cmd}\n')
+    else:
+        print(f'ERROR: path "{path}" does not exist')
